cvd-service-paef-guilherme (1)/cvd-service-paef-guilherme/pyproject.toml
[tool.poetry]
name = "cvd-service-paef"
version = "0.2.0"
description = "Precificação de Ativos de Empresas Fraudulentas"
authors = ["Scarp IT"]

[tool.poetry.dependencies]
python = "^3.9"
flask = {version = "*", extras = ["async"]}
pandas = "*"
gitpython = "*"
sqlalchemy = "*"
psycopg2-binary = "*"
python-dotenv = "*"

[tool.poetry.scripts]
run-app = "src.index:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
